refactor Move and use pregenerated objects.
A move no longer stores the moving piece (see MovingPieceDecorator).
(In the longer term, this info is probably not required at all, since can be retrieved
from the current board.)

Because of this refactoring it's now possible to have constant Move objects, e.g.
for castling, enpassant, and pawn moves (not yet captures). Pawn-Promotion to come...